+++ test/cases/split-chunk-node-module-many-vendors/expected/assets/js/main.719a7fc9.js
@@ -1,0 +1,1 @@
+"use strict";(self.webpackChunk=self.webpackChunk||[]).push([[179],{225:(e,l,o)=>{o.r(l);var s=o(815);console.log(">> main.js"),console.log("Lorem: ",s.lorem.getTitle()),console.log("Lib 1: ",s.libA.getName()),console.log("Lib 2: ",s.libB.getName())}},e=>{e.O(0,[469,621],(()=>(225,e(e.s=225)))),e.O()}]);