+++ test/cases/split-chunk-resolve-assets-dev/src/assets/js/lib.js
@@ -1,0 +1,13 @@
+const Lib = {
+  methodA() {
+    return 'module A';
+  },
+  methodB() {
+    return 'module B';
+  },
+  methodC() {
+    return 'common used module';
+  },
+};
+
+module.exports = Lib;