+++ test/cases/split-chunk-resolve-assets-dev/src/assets/js/main-b.js
@@ -1,0 +1,6 @@
+const modB = require('./module-b');
+const modC = require('./module-c');
+
+console.log('>> main-b:');
+console.log(' - B: ', modB);
+console.log(' - C: ', modC);