+++ test/cases/split-chunk-resolve-assets-dev/src/assets/js/module-c.js
@@ -1,0 +1,4 @@
+const lib = require('./lib');
+const value = lib.methodC();
+
+module.exports = value;