+++ test/cases/split-chunk-resolve-assets-prod/expected/assets/js/168.5a1ae208.js
@@ -1,0 +1,1 @@
+(self.webpackChunk=self.webpackChunk||[]).push([[168,541],{541:(o,l,e)=>{const n=e(416),s=e(703);console.log(">> main-a:"),console.log(" - A: ",n),console.log(" - C: ",s)}}]);