+++ test/cases/split-chunk-resolve-assets-prod/expected/assets/js/416.dfafe31f.js
@@ -1,0 +1,1 @@
+(self.webpackChunk=self.webpackChunk||[]).push([[416],{416:(e,s,h)=>{const k=h(998).methodA();e.exports=k}}]);