+++ test/cases/split-chunk-resolve-assets-prod/expected/assets/js/691.e11602ee.js
@@ -1,0 +1,1 @@
+(self.webpackChunk=self.webpackChunk||[]).push([[691],{691:(e,s,h)=>{const k=h(998).methodB();e.exports=k}}]);