+++ test/cases/split-chunk-resolve-assets-prod/src/assets/js/main-a.js
@@ -1,0 +1,6 @@
+const modA = require('./module-a');
+const modC = require('./module-c');
+
+console.log('>> main-a:');
+console.log(' - A: ', modA);
+console.log(' - C: ', modC);