+++ test/cases/split-chunk-resolve-assets-prod/src/assets/js/module-a.js
@@ -1,0 +1,4 @@
+const lib = require('./lib');
+const value = lib.methodA();
+
+module.exports = value;