+++ test/cases/split-chunk-vendor/expected/assets/js/fixture-lorem.e3b79b01.js
@@ -1,0 +1,1 @@
+"use strict";(self.webpackChunk=self.webpackChunk||[]).push([[890],{},s=>{s.O(0,[216],(()=>(784,s(s.s=784)))),s.O()}]);