+++ test/cases/split-chunk-vendor/expected/assets/js/main.bce9d501.js
@@ -1,0 +1,1 @@
+(self.webpackChunk=self.webpackChunk||[]).push([[179],{225:()=>{console.log(">> main.js")}},s=>{s(s.s=225)}]);